+++ test/integration/tests/button/props.js
@@ -140,6 +140,54 @@ describe(`paypal button component props`, () => {
     );
   });
 
+  it("should render a button and have the referrerDomain in xprops", () => {
+    const expectReferrerDomainToEqual = (uri, domain) => {
+      Object.defineProperty(document, "referrer", {
+        get: () => uri,
+        configurable: true,
+      });
+
+      return wrapPromise(({ expect }) => {
+        window.paypal
+          .Buttons({
+            test: {
+              action: "checkout",
+              onRender: expect("onRender", ({ xprops }) => {
+                if (xprops.referrerDomain !== domain) {
+                  throw new Error(
+                    `Expected referrerDomain to be ${domain || ""}, got ${
+                      xprops.referrerDomain
+                    }`
+                  );
+                }
+              }),
+            },
+          })
+          .render("#testContainer");
+      });
+    };
+
+    return ZalgoPromise.all([
+      expectReferrerDomainToEqual(
+        "https://example.com/path?q=1",
+        "example.com"
+      ),
+      expectReferrerDomainToEqual(
+        "https://not.example.com/path?q=1",
+        "not.example.com"
+      ),
+      expectReferrerDomainToEqual(
+        // eslint-disable-next-line no-script-url
+        "javascript:alert(document.cookie)",
+        undefined
+      ),
+      expectReferrerDomainToEqual(
+        "", // when there is no referrer
+        undefined
+      ),
+    ]);
+  });
+
   it("should render a button and get the renderedButtons props", () => {
     // should not render applepay without applepay listed in xprops.enableFunding
     const renderedButtons = [FUNDING.PAYPAL, FUNDING.APPLEPAY, FUNDING.CARD];